# apps/web/public/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "SlyncPay API",
    "version": "2.0.0",
...
  },
  "servers": [
    { "url": "https://slyncpay-api.onrender.com", "description": "Production" }
  ],
  "security": [{ "bearerAuth": [] }],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "API Key"
      }
    },
    "schemas": {
      "Error": {
        "type": "object",
        "properties": {
          "error": { "type": "string", "example": "validation_error" },
          "message": { "type": "string", "example": "Field xyz is required" },
          "statusCode": { "type": "integer", "example": 400 }
        }
      },
...
      "LineItem": {
        "type": "object",
        "required": ["amountCents"],
        "properties": {
          "description": { "type": "string", "maxLength": 500 },
          "amountCents": { "type": "integer", "minimum": 1 },
          "quantity": { "type": "number", "minimum": 0 },
          "unit": { "type": "string", "maxLength": 50 }
        }
      },
      "Payable": {
        "type": "object",
        "properties": {
          "id": { "type": "string", "format": "uuid" },
          "entityId": { "type": "string", "format": "uuid" },
          "workerId": { "type": "string", "format": "uuid" },
          "engagementId": { "type": "string", "format": "uuid" },
          "externalReferenceId": { "type": ["string", "null"], "description": "Your reference (e.g. invoice or shift ID)." },
          "amountCents": { "type": "integer", "minimum": 1 },
          "status": { "type": "string", "enum": ["draft", "pending", "processing", "paid", "failed", "cancelled"] },
          "dueDate": { "type": "string", "format": "date" },
          "lineItems": { "type": "array", "items": { "$ref": "#/components/schemas/LineItem" } },
          "disbursementId": { "type": ["string", "null"], "format": "uuid", "description": "Set once the payable has been swept into a disbursement batch." },
          "paidAt": { "type": ["string", "null"], "format": "date-time" },
          "createdAt": { "type": "string", "format": "date-time" },
          "updatedAt": { "type": "string", "format": "date-time" }
        }
      },
      "WorkerOnPayable": {
        "type": "object",
        "description": "Worker info embedded in disbursement detail payables.",
        "properties": {
          "id": { "type": "string", "format": "uuid" },
          "firstName": { "type": ["string", "null"] },
          "lastName": { "type": ["string", "null"] },
          "email": { "type": "string", "format": "email" },
          "externalId": { "type": "string" }
        }
      },
      "PayableInDisbursement": {
        "allOf": [
          { "$ref": "#/components/schemas/Payable" },
          {
            "type": "object",
            "properties": {
              "worker": { "anyOf": [{ "$ref": "#/components/schemas/WorkerOnPayable" }, { "type": "null" }] }
            }
          }
        ]
      },
      "Disbursement": {
        "type": "object",
        "properties": {
          "id": { "type": "string", "format": "uuid" },
          "entityId": { "type": "string", "format": "uuid" },
          "status": { "type": "string", "enum": ["processing", "completed", "failed", "partial"] },
          "totalPayablesCount": { "type": "integer" },
          "totalAmountCents": { "type": "integer" },
          "initiatedAt": { "type": "string", "format": "date-time" },
          "completedAt": { "type": ["string", "null"], "format": "date-time" },
          "failureReason": { "type": ["string", "null"] },
          "submittedToProcessor": { "type": "boolean", "description": "True once the batch has been accepted by the payment processor." }
        }
      },
      "DisbursementDetail": {
        "allOf": [
          { "$ref": "#/components/schemas/Disbursement" },
          {
            "type": "object",
            "properties": {
              "payables": { "type": "array", "items": { "$ref": "#/components/schemas/PayableInDisbursement" } }
            }
          }
        ]
      },
...
      "Pagination": {
        "type": "object",
        "properties": {
          "page": { "type": "integer" },
          "limit": { "type": "integer" },
          "total": { "type": "integer" },
          "hasMore": { "type": "boolean" }
        }
      },
...
    }
  },
...
  "paths": {
...
    "/v1/disbursements": {
      "post": {
        "tags": ["Disbursements"],
        "summary": "Trigger a batch disbursement",
        "description": "Pays out all pending payables for the given entity in a single batch.\n\n**Idempotency-Key header required.**",
        "parameters": [{ "name": "Idempotency-Key", "in": "header", "required": true, "schema": { "type": "string" } }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "type": "object", "required": ["entityId"], "properties": { "entityId": { "type": "string", "format": "uuid" } } },
              "example": { "entityId": "9e1d…-…" }
            }
          }
        },
        "responses": {
          "201": { "description": "Disbursement initiated", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Disbursement" } } } },
          "400": { "description": "No pending payables for this entity", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } } }
        }
      },
      "get": {
        "tags": ["Disbursements"],
        "summary": "List disbursements",
        "parameters": [
          { "name": "entityId", "in": "query", "schema": { "type": "string", "format": "uuid" } },
          { "name": "status", "in": "query", "schema": { "type": "string", "enum": ["processing", "completed", "failed", "partial"] } },
          { "name": "page", "in": "query", "schema": { "type": "integer", "default": 1 } },
          { "name": "limit", "in": "query", "schema": { "type": "integer", "default": 20, "maximum": 100 } }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": { "type": "array", "items": { "$ref": "#/components/schemas/Disbursement" } },
                    "pagination": { "$ref": "#/components/schemas/Pagination" }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/v1/disbursements/{id}": {
      "get": {
        "tags": ["Disbursements"],
        "summary": "Get a disbursement (with payables)",
        "description": "Returns the disbursement plus the payables it contains, each enriched with the worker's name and email. While the disbursement is still `processing`, this endpoint refreshes each payable's status from the underlying processor and rolls the disbursement up to `completed`, `partial`, or `failed` once every payable lands in a terminal state.",
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" } }],
        "responses": {
          "200": { "description": "OK", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/DisbursementDetail" } } } },
          "404": { "description": "Not found", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } } }
        }
      }
    },
...
  }
}